fixed delete product

+++ src/main/java/com/example/fiskenatet/scripts/userProfile.js
@@ -2,7 +2,6 @@ $(document).ready(function () {
     var rootURL = 'http://localhost:8091/api';
     var currentUserID = sessionStorage.getItem("currentUser");
     var currentUser;
-    var currentUserProducts;
 
     getUserById();
     
@@ -17,15 +16,29 @@ $(document).ready(function () {
     });
 
     $(document).on("click", "#lnkDeleteProduct", function () {
-        var currentProductId = $(this).data("value");
-        deleteProduct(currentProductId);
-
-        location.href="../webcontent/index.html";
+        checkAuctionForBids();
     });
 
     $(document).on("click", "#btnDeleteUser", function () {
-        deleteUser()
+        checkUserForActiveAuctions();
     });
+
+    function checkAuctionForBids(){
+        var currentProductId = $(this).data("value");
+        var currentProduct = getProductById(currentProductId);
+        if(currentProduct['listOfBids'].length == 0){
+            deleteProduct(currentProductId);
+        }else{
+            alert("Misslyckades, Man kan inte ta bort en annons någon budat på!");
+        }
+    }
+    function checkUserForActiveAuctions() {
+        if(currentUser['listOfProducts'].length == 0){
+            deleteUser();
+        }else{
+            alert("Det går inte att ta bort ett konto med aktiva auktioner!");
+        }
+    }
     
     function deleteUser() {
         $.ajax({
@@ -50,6 +63,19 @@ $(document).ready(function () {
             }
         });
     }
+    function getProductById(currentProductId) {
+        $.ajax({
+            type: 'GET',
+            contentType: 'application/json',
+            url: rootURL + '/products/' + currentProductId,
+            success: function (data, textStatus, jgXHR) {
+                currentUser = data;
+                populateUserInfo();
+                populateUserProducts();
+            }
+        });
+    }
+
 
     function getUserById() {
         $.ajax({
